fix middlew.auth,orFail,delCard,add controll.auth
