+++ controllers/users.js
@@ -1,8 +1,5 @@
-const bcrypt = require('bcrypt');
-const jwt = require('jsonwebtoken');
 const User = require('../models/user');
 const { validationError } = require('./validationError');
-require('dotenv').config();
 
 module.exports.getUsers = (req, res) => {
   User.find({})
@@ -19,26 +16,9 @@ module.exports.getUserId = (req, res) => {
     .orFail(() => {
       res
         .status(404)
-        .send({ message: 'Нет пользователя с таким id!' });
+        .send({ message: `Неправильный ID=${userId} пользователя!` });
     })
     .then((user) => res.send({ data: user }))
-    .catch((err) => validationError(err, req, res));
-};
-
-module.exports.createUser = (req, res) => {
-  const {
-    name, about, avatar, email, password,
-  } = req.body;
-
-  bcrypt.hash(password, 10)
-    .then((hash) => User.create({
-      name, about, avatar, email, password: hash,
-    }))
-    .then((user) => {
-      res
-        .status(201)
-        .send({ data: user });
-    })
     .catch((err) => validationError(err, req, res));
 };
 
@@ -54,7 +34,7 @@ module.exports.updateUser = (req, res) => {
     .orFail(() => {
       res
         .status(404)
-        .send({ message: 'Пользователь не обновляется!' });
+        .send({ message: `Неправильный ID=${owner} пользователя!` });
     })
     .then((user) => res.send({ data: user }))
     .catch((err) => validationError(err, req, res));
@@ -72,33 +52,8 @@ module.exports.updateAvatar = (req, res) => {
     .orFail(() => {
       res
         .status(404)
-        .send({ message: 'Аватар не обновляется!' });
+        .send({ message: `Неправильный ID=${owner} пользователя!` });
     })
     .then((user) => res.send({ data: user }))
     .catch((err) => validationError(err, req, res));
-};
-
-module.exports.login = (req, res) => {
-  const { email, password } = req.body;
-
-  return User.findUserByCredentials(email, password)
-    .then((user) => {
-      const { NODE_ENV, JWT_SECRET } = process.env;
-      const token = jwt.sign(
-        { _id: user._id },
-        NODE_ENV === 'production' ? JWT_SECRET : 'some-secret-key',
-        { expiresIn: '7d' },
-      );
-
-      res
-        .cookie('jwt', token, {
-          maxAge: 3600000 * 24 * 7,
-          httpOnly: true,
-        // sameSite: true,
-        })
-        .end();
-    })
-    .catch((err) => res
-      .status(401)
-      .send({ message: err.message }));
 };
