+++ routes/users.js
@@ -1,10 +1,8 @@
 const router = require('express').Router();
 const {
-  getUsers, getUserId, createUser, updateUser, updateAvatar, login,
+  getUsers, getUserId, updateUser, updateAvatar,
 } = require('../controllers/users');
 
-router.post('/', createUser);
-router.post('/', login);
 router.get('/', getUsers);
 router.get('/:userId', getUserId);
 router.patch('/me', updateUser);
